Add trained model states
