+++ recognition/VQVAE_44310776/.gitignore
@@ -1,5 +1,4 @@
 data
-results
 .ipynb_checkpoints
 __pycache__/
 .env
